avoid to rebuild sut

+++ build_all_binaries.sh
@@ -1,6 +1,9 @@
 #! /bin/bash
 
 for project in ./sut-images/* ; do
+
+
+
     project_path=$(realpath $project)
     echo $project_path
     project_name=$(basename $project_path)
@@ -13,6 +16,10 @@ for project in ./sut-images/* ; do
         echo $type
         echo $cmd
         build_dir=$HOME/sut-docker/$project_name/$type/
+        if [ -d "$build_dir" ]; then
+            echo "$build_dir does exist."
+            continue
+        fi
         mkdir -p $build_dir
         docker run -it --rm -v $build_dir:/out -t fuzzbuild/$project_name $cmd
     done
